api_gateway/app/pyproject.toml
[tool.poetry]
name = "api-gateway"
version = "0.1.0"
description = ""
authors = ["kpimenov <kirill.pimenoff@mail.ru>"]
readme = "README.md"
packages = [{include = "api_gateway"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
aiohttp = "^3.8.4"
uvicorn = "^0.21.1"
python-jose = "^3.3.0"
pydantic = {extras = ["email"], version = "^1.10.7"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
